Create swgdbr.py
Committing this in "beta" form. It works in a minimal way to start a beacon that randomly changes after some chosen time in minutes.
Usage from cli is 
  >sudo swgedbr.py x &
x is minutes. If run with 0, it will stop a previous run of the command and disable ble adv.
There are several known issues, including:
- If the process is not shutdown with "sudo swgedbr.py 0" or SIGTERM, the /var/swgedbr.pid file will have to be removed manually (>sudo rm -f /var/swgedbr.pid)
- Calling the command a second time with x>0 will cause an exception. The .pid file will be written with binary values and fail
I'm sure there are other failings but I haven't fully tested yet.
This commit is in preparation of changes to the CherryPy web script to enable a timed random beacon change functionality
Once that is done, I'll come back to this with error handling, cli usage, and other possible usage fixes. In it's current form it will satisfy the needs for the web UI