Create swgedbr.py
had to recreate bc I typo'd the file name
provides timed random beacon changes through cli or the Web UI
recommend to copy to the /opt directory. This file name and path are currently hard coded to the web UI and that will need to be updated if you decide to do differently.